+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 * :feature:`-` Current and historical token prices will now also be queryable via Alchemy.
 * :feature:`9083` rotki will now correctly decode ParaSwap v6 swaps in all supported EVM chains.
 * :feature:`-` Cowswap in base is now supported.
+* :feature:`9270` rotki will now support Echo.xyz funding and refunding event.
 * :feature:`9120` rotki will now correctly decode Firebird Finance swaps in all supported EVM chains.
 * :feature:`7254` rotki will now correctly decode OpenOcean swaps in all supported EVM chains.
 * :feature:`-` Lido and EtherFI reserve tokens will now be automatically queried.